examples/fastapi/requirements.txt
aiosqlite==0.17.0
anyio==3.7.0
asgi-lifespan==2.1.0
asyncpg==0.27.0
certifi==2023.5.7
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
iso8601==1.1.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.8
pypika-tortoise==0.1.6
pytest==7.3.1
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
tortoise-orm==0.19.3
typing-extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
zipp==3.15.0
